refactor(media_control_plugin): move Media Control to package Control
