+++ clappr/src/main/kotlin/io/clappr/player/plugin/PluginConfig.kt
@@ -1,5 +1,7 @@
 package io.clappr.player.plugin
 
+import io.clappr.player.plugin.Control.MediaControl
+
 object PluginConfig {
     fun register() {
         // TODO - Add default plugins
